Added an "Existing Page Import" Feature
This creates a unique page identifier that is based on the user-created layout. Within the theme is a new pop-up modal that gives the user the corresponding code for their current page layout. They can copy and paste this code in Page Builder to have Page Builder automatically re-create their layout. This is beneficial for making quick edits to their existing design.