Added Import Layout Mechanisms
Added a mechanism to import a preexisting layout. Through a pop-up modal, It takes in a CSV string, converts into double array (for counts), then converts into objects with "id" and "selectedCustomization" properties. After this, the system begins assembling elements and adding a count User Interface input for such. The elements are generated. Current work to be done mainly revolves around (1) ensuring that all existing elements are removed prior to the user adding their own elements (write separate function for this) and (2) getting multi-variable sections such as pureslider and carousel to actually generate content instances in accordance with .selectedCustomization (currently they do not). In the future look into externalizing JS functions to improve readability of code.